Add docstrings to functions in packages
