+++ nesta/packages/spacy_utils/preprocess_spacy.py
@@ -3,18 +3,45 @@ nlp = spacy.load('en', disable=['ner'])
 
 def convert(text):
 
+    """Function to remove/replace curly quotes (“, ”, ‘, ’)
+
+    Args:
+        text (str): String of document.
+
+    Returns:
+        transformed_doc (str): Curly quotes replaced/removed in document.
+
+    """
+
     return str(text).replace('“',"").replace('”',"").replace('‘',"'").replace('’',"'")
 
-def spacy_nlp_vocab_update(stop_word_list): #put in main function or leave it here?
+def spacy_nlp_vocab_update(stop_word_list):
+    """Function to update Spacy's NLP vocabulary
+
+    Args:
+        stop_word_list (list): List of stop-words.
+
+    """
+
     for word in stop_word_list:
-    l = word.lower()
-    u = word.upper()
-    t = word.title()
-    nlp.vocab[l].is_stop = True
-    nlp.vocab[u].is_stop = True
-    nlp.vocab[t].is_stop = True
+        l = word.lower()
+        u = word.upper()
+        t = word.title()
+        nlp.vocab[l].is_stop = True
+        nlp.vocab[u].is_stop = True
+        nlp.vocab[t].is_stop = True
 
 def word_tokenise(text):
+    """Preprocessing text using spacy library
+
+    Args:
+        text (str): String of document.
+
+    Returns:
+        document_tokens (list): List of the processed document's unigram tokens.
+
+    """
+
     text_ = convert(text)
 
     doc = nlp(text_)
